fix: almost service for corresponding behavior

+++ services/groupService.js
@@ -2,24 +2,22 @@ const BasicService = require('../utils/services/basicService');
 const bindMethodsWithThisContext = require('../utils/classes/bindMethodsWithThisContext');
 const Group = require('../models/group');
 const GroupMember = require('../models/groupMember');
-const { BadRequestException, TargetNotExistException, IncorrectPermission } = require('../utils/exceptions/commonExceptions');
+const { BadRequestException, CommonException, TargetNotExistException, IncorrectPermission, TargetAlreadyExistException } = require('../utils/exceptions/commonExceptions');
 const { ROLES } = require('../utils/constants/users');
 const { updateObjectIfUpdateFieldDataDefined } = require('../utils/objects');
-
 class GroupService extends BasicService {
     constructor() {
         super();
         bindMethodsWithThisContext(this);
     }
     async getGroupWhereUserIsMemberWithPaginate({ userId, page }) {
-        const query = {
-            user: userId,
-            accepted: true,
-            banned: false
-        };
         const paginatedResults = await this.getPaginatedResults({
             model: GroupMember,
-            query,
+            query: {
+                user: userId,
+                accepted: true,
+                banned: false
+            },
             page,
             sort: { createdAt: -1 }
         });
@@ -33,18 +31,31 @@ class GroupService extends BasicService {
         const paginatedResults = await this.getPaginatedResults({
             model: Group,
             query: {
-                $or: [{ admin: userId }, { mod: userId }]
+                $or: [
+                    { admin: userId },
+                    { mod: userId }
+                ]
             },
             page,
         });
 
         return paginatedResults;
     }
-    async getJoinGroupRequestWithPaginate({ groupId, currentUser }) {
+    async getJoinGroupRequestWithPaginate({ groupId, page, currentUser }) {
         const group = await Group.findById(groupId);
         if (!group) {
             throw new BadRequestException('Group not exist');
         }
+        const result = await this.getPaginatedResults({
+            model: GroupMember,
+            query: {
+                group: group._id,
+                accepted: false,
+                banned: false
+            },
+            page
+        });
+        return result;
     }
     async createGroup({ currentUser, name, description, privacy, needApprovedToJoin }) {
         const group = new Group({
@@ -58,16 +69,16 @@ class GroupService extends BasicService {
         return await group.save();
     }
     checkRoleOfUserInGroup({ userId, group }) {
-        if (group.admin.contains(userId)) {
+        if (group.admin.find(x => x.equals(userId))) {
             return ROLES.ADMIN;
         }
-        if (group.mod.contains(userId)) {
+        if (group.mod.find(x => x.equals(userId))) {
             return ROLES.MOD;
         }
         return null;
     }
     checkRoleOfCurrentUserInGroup({ currentUser, group }) {
-        if (currentUser.role != ROLES.USER) {
+        if (currentUser.role !== ROLES.USER) {
             return currentUser.role;
         }
         return this.checkRoleOfUserInGroup({
@@ -77,121 +88,212 @@ class GroupService extends BasicService {
     }
     checkCurrentUserCanModifyGroup({ currentUser, group }) {
         const role = this.checkRoleOfCurrentUserInGroup({ currentUser, group });
-        if (!role) {
+        if (role !== ROLES.ADMIN && role !== ROLES.MOD) {
             throw new IncorrectPermission('You do not have permission');
         }
     }
-    async updateGroup({ currentUser, groupId, name, description, avatar, cover, location, privacy, needApprovedToJoin }) {
+    async updateGroup({ currentUser, groupId, name, description, avatar, cover, location, privacy, needApprovedToJoin, needApprovedToPost }) {
         const group = await Group.findById(groupId);
         if (!group) {
-            throw new BadRequestException('Group not exist');
+            throw new TargetNotExistException('Group not exist');
         }
         this.checkCurrentUserCanModifyGroup({
             currentUser,
             group
         });
-        updateObjectIfUpdateFieldDataDefined(group, { groupId, name, description, avatar, cover, location, privacy, needApprovedToJoin });
+        updateObjectIfUpdateFieldDataDefined(group, { groupId, name, description, avatar, cover, location, privacy, needApprovedToJoin, needApprovedToPost });
+        //TODO: push notification for all member of group including other admin mod
         return await group.save();
     }
     async deleteGroup({ id, currentUser }) {
         const group = await Group.findById(id);
+        if (!group) {
+            throw new TargetNotExistException('Group not exist');
+        }
         const role = this.checkRoleOfCurrentUserInGroup({ currentUser, group });
-        if (role != ROLES.ADMIN) {
+        if (role !== ROLES.ADMIN) {
             throw new IncorrectPermission('You do not have permission');
         }
         //TODO: must delete all post in group also
+        //Push notification to all member
         await GroupMember.deleteMany({
             group: group._id
         });
-        await group.remove();
+        await Group.findByIdAndDelete(group._id);
     }
     async joinGroup({ id, currentUser }) {
         const group = await Group.findById(id);
         if (!group) {
-            throw new BadRequestException('Group not exist');
+            throw new TargetNotExistException('Group not exist');
         }
         const existRole = this.checkRoleOfUserInGroup({
             group,
             userId: currentUser.userId
         });
-        if (existRole == ROLES.ADMIN || existRole == ROLES.MOD) {
-            throw new BadRequestException('You are already in this group');
+        if (existRole === ROLES.ADMIN || existRole === ROLES.MOD) {
+            throw new TargetAlreadyExistException('You are already in this group');
+        }
+        const existJoinRequest = await GroupMember.findOne({
+            group: group._id,
+            user: currentUser.userId,
+        });
+
+        if (existJoinRequest) {
+            if (existJoinRequest.accepted) {
+                throw new TargetAlreadyExistException('You have already in this group');
+            }
+            throw new TargetAlreadyExistException('You have sent join request, you do not need to send it again');
         }
         const groupMember = new GroupMember({
             group: group._id,
             user: currentUser.userId,
             accepted: !group.needApprovedToJoin
         });
+
         try {
             await groupMember.save();
+            //PUSH notification for admin/mod
         }
         catch (err) {
             console.log(err);
             throw new BadRequestException('Something went wrong or you has been banned from the group');
         }
     }
+    async acceptJoinRequest({ id, currentUser }) {
+        const request = await GroupMember.findById(id);
 
-    async banUserFromGroup({ currentUser, userId, groupId }) {
+        if (!request) {
+            throw new TargetNotExistException('Request not exist');
+        }
+        if (request.banned) {
+            throw new IncorrectPermission('The user have been banned of this group');
+        }
+        if (request.accepted) {
+            throw new BadRequestException('Request already accepted');
+        }
+        const group = await Group.findById(request.group);
+        if (!group) {
+            await GroupMember.deleteMany({ group: request.group });
+            throw new CommonException('The request is invalid');
+        }
+        this.checkCurrentUserCanModifyGroup({ currentUser, group });
+
+        request.accepted = true;
+        await request.save();
+        //TODO: push notification to the user
+    }
+    async rejectJoinRequest({ id, currentUser }) {
+        const request = await GroupMember.findById(id);
+
+        if (!request) {
+            throw new TargetNotExistException('Request not exist');
+        }
+        if (request.banned) {
+            throw new IncorrectPermission('The user have been banned of this group');
+        }
+        if (request.accepted) {
+            throw new BadRequestException('Request already accepted');
+        }
+        const group = await Group.findById(request.group);
+        if (!group) {
+            await GroupMember.deleteMany({ group: request.group });
+            throw new CommonException('The request is invalid');
+        }
+        this.checkCurrentUserCanModifyGroup({ currentUser, group });
+
+        await GroupMember.findByIdAndDelete(request._id);
+        //TODO: push notification to the user
+    }
+    async getAllUserInGroup({ currentUser, page, groupId }) {
+        const group = await Group.findById(groupId);
+        if (!group) {
+            throw new BadRequestException('Group not exist');
+        }
+        const members = await this.getPaginatedResults({
+            model: GroupMember,
+            query: {
+                group: group._id,
+                accepted: true,
+                banned: false,
+            },
+            page,
+        });
+        const result = {
+            ...group.toObject(),
+            members,
+        };
+
+        return result;
+    }
+    async getAllBannedUserInGroup({ currentUser, page, groupId }) {
+        const group = await Group.findById(groupId);
+        if (!group) {
+            throw new BadRequestException('Group not exist');
+        }
+        const members = await this.getPaginatedResults({
+            model: GroupMember,
+            query: {
+                group: group._id,
+                banned: true,
+            },
+            page,
+        });
+        const result = {
+            ...group.toObject(),
+            members,
+        };
+
+        return result;
+    }
+    async toggleBanUserFromGroup({ currentUser, userId, groupId, banned = true }) {
         const group = await Group.findById(groupId);
         if (!group) {
             throw new BadRequestException('Group not exist');
         }
         this.checkCurrentUserCanModifyGroup({ currentUser, group });
 
-        const groupMember = GroupMember.findOne({
+        const groupMember = await GroupMember.findOne({
             group: group._id,
             user: userId,
-            banned: false,
         });
         if (!groupMember) {
-            throw new BadRequestException('User does not in the group or has been banned');
+            throw new BadRequestException('User does not in the group');
         }
-        groupMember.banned = true;
+        if (groupMember.banned == banned) {
+            throw new TargetAlreadyExistException('The user already have been ' + (banned ? 'banned' : 'un banned'));
+        }
+        groupMember.banned = banned;
         await groupMember.save();
-    }
-    async leaveGroup({ id, currentUser }) {
-        const group = await Group.findById(id);
-        if (!group) {
-            throw new BadRequestException('Group does not exist');
-        }
-    
-        const role = this.checkRoleOfCurrentUserInGroup({ currentUser, group });
-    
-        if (role) {
-            await this.removeUserInGroup({
-                groupId: id,
-                userId: currentUser.userId
-            });
-        } else {
-            await this.removeMemberInGroup({
-                userId: currentUser.userId,
-                groupId: id
-            });
-        }
     }
     async appointUserForGroup({ id, userId, newRole, currentUser }) {
         //TODO: check userId exist first
         const group = await Group.findById(id);
         const currentUserRole = this.checkRoleOfCurrentUserInGroup({ currentUser, group });
-        if (currentUserRole != ROLES.ADMIN) {
+        if (currentUserRole !== ROLES.ADMIN) {
             throw new IncorrectPermission('You do not have permission');
         }
-        if (newRole = ROLES.ADMIN) {
+        if (newRole === ROLES.ADMIN) {
             group.admin.push(userId);
         }
-        else if (newRole = ROLES.MOD) {
+        else if (newRole === ROLES.MOD) {
             group.mod.push(userId);
         }
         else {
             throw new BadRequestException('Invalid role specified');
         }
         await group.save();
+        //TODO: push notification to the user
     }
-    async removeAdminOrModInGroup({ userId, group }) {
+    async removeAdminOrModInGroup({ userId, currentUser, group }) {
+        const currentUserRoleOfTheGroup = this.checkRoleOfUserInGroup({ userId: currentUser.userId, group });
+
+        if (currentUserRoleOfTheGroup !== ROLES.ADMIN) {
+            throw new IncorrectPermission();
+        }
         const role = this.checkRoleOfUserInGroup({ userId, group });
-    
+
         if (role === ROLES.ADMIN) {
-            if (group.admin.length === 1) {
+            if (group.admin.length == 1) {
                 throw new BadRequestException('You are the only admin, assign another admin before leaving');
             }
             group.admin = group.admin.filter(id => !id.equals(userId));
@@ -200,32 +302,38 @@ class GroupService extends BasicService {
         } else {
             throw new BadRequestException('User is not an admin or mod in the group');
         }
-    
         await group.save();
+        //TODO: push notification for the user
     }
-    async removeMemberInGroup({ userId, currentUser, groupId }) {
+    async removeMemberInGroup({ userId, currentUser, group }) {
+        const currentUserRoleOfTheGroup = this.checkRoleOfUserInGroup({ userId: currentUser.userId, group });
+
+        if (userId != currentUser.userId && currentUserRoleOfTheGroup !== ROLES.ADMIN && currentUserRoleOfTheGroup !== ROLES.MOD) {
+            throw new IncorrectPermission();
+        }
         const groupMember = await GroupMember.findOneAndDelete({
             user: userId,
-            group: groupId,
+            group: group._id,
         });
         if (!groupMember) {
-            throw new TargetNotExistException('You are not in the group');
+            throw new TargetNotExistException('The user are not in the group');
         }
+        //TODO: push notification to the user
         return groupMember;
     }
-    async removeUserInGroup({ groupId, userId }) {
+    async removeUserInGroup({ groupId, currentUser, userId }) {
         const group = await Group.findById(groupId);
         if (!group) {
             throw new TargetNotExistException('Group does not exist');
         }
-    
+
         const role = this.checkRoleOfUserInGroup({ userId, group });
-    
-        if (role) {
-            await this.removeAdminOrModInGroup({ userId, group });
+        if (role === ROLES.ADMIN || role === ROLES.MOD) {
+            await this.removeAdminOrModInGroup({ userId, currentUser, group });
         }
-    
-        await this.removeMemberInGroup({ userId, groupId });
+        else {
+            await this.removeMemberInGroup({ userId, currentUser, group });
+        }
     }
 }
